feat: Visualise obstacle grid generation
The creation of obstacle grid isn't accompanied by a visual. On grid
creation, its image gets saved locally. It would be nice to see dynamic
updates on the grid as and when it gets populated. Moreover, this feature
is a must have when we go on to implement different graph search algos.
on the grid.

This practically eliminates the need to save images to disk. However, if
required one can pass the command line arg- '--save=True' when running
the script.

Fixes https://github.com/radhasaraf/RBE550-Motion-planning/issues/1