feat: Make display of grid optional using kwargs
When importing the module to create an obstacle grid, one might
not be interested in viewing the formation of the grid and would
most likely be looking for a function that gives you the final o/p
to maybe try and run some planning algos on.

With that consideration in mind, the display of the grid has been
made optional with this commit.

It can be turned on by passing the boolean kwarg 'display' to the
'create_obstacle_grid' function or using the command line argument
'display' when running the module as a script.