maint: Establish practise of 'requirements.txt' for each HW
If someone or somewhere down the line I myself am required to
run a particular homework again, I shouldn't be bothered to
install dependencies for other homeworks which is the case at
present because of a common requirements.txt. Hence, with this
commit, we establish a practise of each homework having its own
requirements.txt file.