refactor: Update dijkstra's logic for diagonal motion
So far, Dijkstra's was exactly similar to BFS since the
neighbours of a given node didn't include the diagonally
adjacent nodes which meant equal weights for all neighbors
at the same level.

This was changed here by modifying the graph representation
of the grid for Dijkstra's by including the diagonal neighbors
and associating a cost with each neighbour equal to its dist
from the node.