fix: Prevent glitch in plotting traversals for BFS and Dijkstras
A correct graph traversal would start at the nodes adjacent to
the start node and end at the end node. Plotting the graph traversal
for BFS and Dijkstras was sometimes incorrect in that it would
start from some random level and then later make up for the missed
levels. (Nodes explored at the same time are all on the same level.)

It turned out this was because the dictionary that stored the
nodes against the corresponding levels may not always be sorted
by levels. Thus, sorting the dictionary elements prior to iterating
through them did the job.